requirements.txt
# YOLOv2-PyTorch Requirements
# Python >= 3.7

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Data Processing
numpy>=1.20.0
PyYAML>=5.4.0

# Training & Evaluation
tqdm>=4.60.0
matplotlib>=3.3.0

# Optional: Visualization
# tensorboard>=2.8.0
# wandb>=0.12.0

# Optional: Performance
# albumentations>=1.0.0  # Advanced augmentations
# pycocotools>=2.0.0  # COCO evaluation

# Development
# pytest>=6.2.0
# black>=21.0
# flake8>=3.9.0
